fixing negative event numbers. Details: https://its.cern.ch/jira/browse/ATLASVPONE-287. (VP1Base-00-23-12)
    * fixing negative event numbers.
	  Details: https://its.cern.ch/jira/browse/ATLASVPONE-287.
	* new tag: VP1Base-00-23-12


Former-commit-id: 9a6cf6caaa3dadb7721ab80fdf5673b893393ae9

+++ graphics/VP1/VP1Base/src/IVP1ChannelWidget.cxx
@@ -38,7 +38,7 @@ public:
 	bool canregistercontroller;
 	bool canregistersystem;
 	int runnumber;
-	int eventnumber;
+	unsigned long long eventnumber;
 	unsigned timestamp;
 };
 
@@ -86,7 +86,7 @@ bool IVP1ChannelWidget::isAccumulator() const
 }
 
 //_______________________________________________________
-void IVP1ChannelWidget::setRunEvtNumber(int runnumber, int eventnumber)
+void IVP1ChannelWidget::setRunEvtNumber(int runnumber, unsigned long long eventnumber)
 {
 	d->runnumber = runnumber;
 	d->eventnumber = eventnumber;
@@ -252,7 +252,7 @@ void IVP1ChannelWidget::turnOff(IVP1System*s,const bool& immediateErase)
 }
 
 //_______________________________________________________
-void IVP1ChannelWidget::getRunEvtNumber(int& runnumber, int& eventnumber)
+void IVP1ChannelWidget::getRunEvtNumber(int& runnumber, unsigned long long& eventnumber)
 {
 	runnumber = d->runnumber;
 	eventnumber = d->eventnumber;
